deploy/requirements.txt
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
streamlit==1.38.0
ipykernel==6.29.5
python-dotenv==1.0.1
st-pages==1.0.1
wikipedia==1.4.0
neo4j==5.20.0
langchain==0.2.0
langchain_google_vertexai==1.0.4
langchain_community==0.2.0
langchain_experimental==0.0.60
google-cloud-aiplatform==1.51.0
yfiles_jupyter_graphs==1.8.1
langchainhub==0.1.21
chardet==5.2.0